Move uniqueId logic to utils so it can be used in other transpilers
